refactor(desktop): finish schema-driven settings
